# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  database.Partner:
    properties:
      companyName:
        type: string
      createdAt:
        type: string
      id:
        type: string
      metaData:
        additionalProperties: true
        type: object
      parent:
        $ref: '#/definitions/database.Partner'
      parentId:
        type: string
      referralCode:
        type: string
      socialMedia:
        additionalProperties: true
        description: |-
          SocialMedia mirrors PartnerApplication.SocialMedia at approve time.
          Application stays as the review snapshot; this column is the
          partner's *current* profile (Settings edits land here).

          `serializer:json` round-trips `map[string]interface{}` through
          encoding/json so the column can stay TEXT — works with both lib/pq
          and pgx. Partial `Updates(map)` bypass the serializer, so the
          service marshals the value to a JSON string before assigning.
        type: object
      status:
        $ref: '#/definitions/database.PartnerStatus'
      tier:
        $ref: '#/definitions/database.PartnerTier'
      tierId:
        type: string
      totalEarnings:
        type: number
      totalReferrals:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/database.User'
      userId:
        type: string
      website:
        type: string
    type: object
...
  database.PartnerStatus:
    enum:
    - pending
    - active
    - suspended
    type: string
    x-enum-varnames:
    - PartnerStatusPending
    - PartnerStatusActive
    - PartnerStatusSuspended
  database.PartnerTier:
    properties:
      color:
        type: string
      commissionRate:
        type: number
      createdAt:
        type: string
      id:
        type: string
      isDefault:
        type: boolean
      level:
        type: integer
      maxVolume:
        type: number
      minActiveClients:
        type: integer
      minVolume:
        type: number
      name:
        type: string
      updatedAt:
        type: string
    type: object
  database.Payout:
    properties:
      amount:
        type: number
      approvedBy:
        type: string
      commissionCount:
        type: integer
      createdAt:
        type: string
      currency:
        type: string
      failureReason:
        type: string
      id:
        type: string
      partner:
        $ref: '#/definitions/database.Partner'
      partnerId:
        type: string
      periodEnd:
        type: string
      periodStart:
        type: string
      processedAt:
        type: string
      status:
        $ref: '#/definitions/database.PayoutStatus'
      transactionId:
        type: string
      updatedAt:
        type: string
    type: object
  database.PayoutStatus:
    enum:
    - pending
    - processing
    - completed
    - failed
    type: string
    x-enum-varnames:
    - PayoutStatusPending
    - PayoutStatusProcessing
    - PayoutStatusCompleted
    - PayoutStatusFailed
...
  database.User:
    properties:
      binanceEmail:
        type: string
      canTrade:
        type: boolean
      canWithdraw:
        type: boolean
      createdAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      isWhitelistEnabled:
        type: boolean
      kycLevel:
        type: integer
      lastName:
        type: string
      metaData:
        additionalProperties: true
        type: object
      status:
        type: integer
      subAccountId:
        type: string
      updatedAt:
        type: string
      vipLevel:
        type: integer
    type: object
...
  structs.ErrorResponse:
    properties:
      body:
        type: string
      error_msg:
        type: string
      status_code:
        type: integer
    type: object
...
  structs.PaginationResponse:
    properties:
      items: {}
      total:
        type: integer
    type: object
...
  structs.PayoutListParams:
    properties:
      current:
        type: integer
      export:
        type: boolean
      limit:
        description: alias for PageSize
        type: integer
      page:
        description: alias for Current
        type: integer
      pageSize:
        type: integer
      partnerId:
        type: string
      sortDate:
        $ref: '#/definitions/structs.SortDate'
      status:
        type: string
    type: object
  structs.PayoutReviewParams:
    properties:
      failureReason:
        type: string
    type: object
...
  structs.ResponseBody:
    properties:
      body: {}
      message:
        type: string
    type: object
  structs.SortDate:
    properties:
      end_day:
        type: string
      start_day:
        type: string
    type: object
...
host: localhost:8090
info:
  contact: {}
  description: |-
    Partner program backend — public, partner, admin, system endpoints.
    Authentication is split into three layers:
    • BearerAuth (PartnerAuth/AdminAuth): Cognito ID token in `Authorization: Bearer …`.
    • InternalKey: `X-Internal-Key` header — only the xmeta-monorepo server should send this.
  title: xmeta Partner API
  version: "1.0"
paths:
...
  /admin/partner/payouts/{id}/approve:
    post:
      consumes:
      - application/json
      description: Approves a pending payout for processing
      parameters:
      - description: Payout ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/structs.ResponseBody'
            - properties:
                body:
                  $ref: '#/definitions/database.Payout'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Approve payout
      tags:
      - Admin Payouts
  /admin/partner/payouts/{id}/reject:
    post:
      consumes:
      - application/json
      description: Rejects a pending payout with a reason
      parameters:
      - description: Payout ID
        in: path
        name: id
        required: true
        type: string
      - description: Rejection reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.PayoutReviewParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/structs.ResponseBody'
            - properties:
                body:
                  $ref: '#/definitions/database.Payout'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reject payout
      tags:
      - Admin Payouts
  /admin/partner/payouts/detail/{id}:
    get:
      consumes:
      - application/json
      description: Returns one payout with line items for admin review
      parameters:
      - description: Payout ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/structs.ResponseBody'
            - properties:
                body:
                  $ref: '#/definitions/database.Payout'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get payout detail (admin)
      tags:
      - Admin Payouts
  /admin/partner/payouts/list:
    post:
      consumes:
      - application/json
      description: Returns a paginated list of payout records across all partners
      parameters:
      - description: Filters and pagination
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.PayoutListParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/structs.ResponseBody'
            - properties:
                body:
                  $ref: '#/definitions/structs.PaginationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List payouts (admin)
      tags:
      - Admin Payouts
  /admin/partner/payouts/pending:
    post:
      consumes:
      - application/json
      description: Returns payouts awaiting admin approval, paginated
      parameters:
      - description: Filters and pagination
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.PayoutListParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/structs.ResponseBody'
            - properties:
                body:
                  $ref: '#/definitions/structs.PaginationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List pending payouts
      tags:
      - Admin Payouts
...
  /partner/payouts/detail/{id}:
    get:
      consumes:
      - application/json
      description: Returns one payout record with its line items
      parameters:
      - description: Payout ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/structs.ResponseBody'
            - properties:
                body:
                  $ref: '#/definitions/database.Payout'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get payout detail
      tags:
      - Partner Payouts
  /partner/payouts/list:
    post:
      consumes:
      - application/json
      description: Returns a paginated list of payout records for the partner
      parameters:
      - description: Filters and pagination
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.PayoutListParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/structs.ResponseBody'
            - properties:
                body:
                  $ref: '#/definitions/structs.PaginationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List partner payouts
      tags:
      - Partner Payouts
  /partner/payouts/pending:
    get:
      consumes:
      - application/json
      description: Returns the partner's currently pending payout balance and breakdown
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.ResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Pending payout balance
      tags:
      - Partner Payouts
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: '"Bearer {Cognito ID Token}" format'
    in: header
    name: Authorization
    type: apiKey
  InternalKey:
    description: Server-to-server API key shared with xmeta-monorepo
    in: header
    name: X-Internal-Key
    type: apiKey
swagger: "2.0"